requirements.txt
# Core Dependencies
numpy>=1.26.0
pillow>=10.0.0
opencv-python>=4.8.0
python-dotenv>=1.0.0

# Deep Learning
ultralytics>=8.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# API
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
